Finish testing Firebase signup testing
Fixed a few logical error + successfully registered user for Firebase
