+++ app/src/main/res/layout/activity_main.xml
@@ -1,18 +1,93 @@
 <?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    android:background="@color/black"
+    android:paddingTop="200dp"
 
-    <TextView
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="Hello World!"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
+    tools:context=".LogInActivity">
 
-</androidx.constraintlayout.widget.ConstraintLayout>
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:orientation="vertical"
+        android:paddingLeft="10dp"
+        android:paddingRight="10dp"
+        android:id="@+id/layout1"
+        android:layout_height="wrap_content">
+        <TextView
+            android:layout_width="match_parent"
+            android:text="Email"
+            android:gravity="center_vertical"
+            android:paddingLeft="20dp"
+            android:textColor="@color/white"
+            android:textSize="20dp"
+            android:layout_height="50dp">
+        </TextView>
+        <EditText
+            android:layout_width="match_parent"
+            android:id="@+id/emailtextview"
+            android:background="@drawable/rectanglebox"
+            android:layout_height="35dp">
+
+        </EditText>
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:orientation="vertical"
+        android:paddingLeft="10dp"
+        android:paddingRight="10dp"
+        android:layout_below="@id/layout1"
+        android:id="@+id/layout2"
+        android:layout_height="wrap_content">
+        <TextView
+            android:layout_width="match_parent"
+            android:text="Password"
+            android:gravity="center_vertical"
+            android:paddingLeft="20dp"
+            android:textColor="@color/white"
+            android:textSize="20dp"
+            android:layout_height="50dp">
+        </TextView>
+        <EditText
+            android:layout_width="match_parent"
+            android:background="@drawable/rectanglebox"
+            android:id="@+id/passwordtextview"
+            android:layout_height="35dp">
+
+        </EditText>
+    </LinearLayout>
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_below="@id/layout2"
+        android:layout_marginTop="10dp"
+        android:gravity="center"
+        android:id="@+id/LoginButtonLayout"
+        android:layout_height="40dp">
+        <Button
+            android:layout_width="200dp"
+            android:text="Login"
+            android:id="@+id/LoginButton"
+            android:layout_height="match_parent">
+        </Button>
+    </LinearLayout>
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_below="@id/LoginButtonLayout"
+        android:layout_marginTop="10dp"
+        android:gravity="center"
+        android:layout_height="40dp">
+        <Button
+            android:layout_width="200dp"
+            android:text="Register"
+            android:id="@+id/RegisterButton"
+            android:layout_height="match_parent">
+        </Button>
+    </LinearLayout>
+
+
+</RelativeLayout>
